pyproject.toml
[tool.poetry]
name = "replay"
version = "0.1.0"
description = "Makes a movie from a photo library"
authors = ["Arvid Larsson <arvla626@student.liu.se>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
scandir = "^1.10.0"
click = "^7.1.1"
regex = "^2020.4.4"
ffmpeg-python = "^0.2.0"
pillow = "^7.1.1"
pandas = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
